+++ test/builtinDetectors.spec.ts
@@ -8,7 +8,6 @@ import {
   getFilePathArg,
 } from "./testUtil";
 import { executeMisti } from "../src/cli";
-import JSONbig from "json-bigint";
 import fs from "fs";
 import path from "path";
 
@@ -71,48 +70,3 @@ if (filePathArg) {
     processProjectDir(projectDir);
   });
 }
-
-describe("JSON output test for a single file", () => {
-  it("should generate valid JSON output for never-accessed-1.tact", async () => {
-    const filePath = path.resolve(__dirname, "good", "never-accessed-1.tact");
-
-    const output = await executeMisti([
-      "--all-detectors",
-      "--no-colors",
-      "--output-format",
-      "json",
-      filePath,
-    ]);
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    let jsonOutput: any;
-    try {
-      jsonOutput = JSONbig.parse(output);
-    } catch (error) {
-      console.error("Bad output:\n", output);
-      throw error;
-    }
-
-    expect(jsonOutput.warnings.length).toBeGreaterThan(0);
-
-    const firstWarning = JSONbig.parse(jsonOutput.warnings[0].warnings[0]);
-    expect(firstWarning).toMatchObject({
-      file: expect.stringContaining("never-accessed-1.tact"),
-      line: expect.any(Number),
-      col: expect.any(Number),
-      detectorId: "NeverAccessedVariables",
-      severity: "MEDIUM",
-      message: expect.stringContaining("Write-only variable: a"),
-    });
-
-    expect(firstWarning.message).toContain(
-      "test/good/never-accessed-1.tact:2:5:",
-    );
-    expect(firstWarning.message).toContain(
-      "Help: The variable value should be accessed",
-    );
-    expect(firstWarning.message).toContain(
-      "See: https://nowarp.io/tools/misti/docs/detectors/NeverAccessedVariables",
-    );
-  });
-});